Se crean algunas consultas de Moderador y Favorito, y se unen los dataset

+++ persistencia/src/main/java/co/edu/uniquindio/uniMarket/repositorios/Detalle_CompraRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Detalle_CompraRepo extends JpaRepository<Detalle_Compra, Long> {
+public interface Detalle_CompraRepo extends JpaRepository<Detalle_Compra, Integer> {
 }
