+++ persistencia/src/main/java/co/edu/uniquindio/uniMarket/repositorios/FavoritoRepo.java
@@ -1,8 +1,13 @@
 package co.edu.uniquindio.uniMarket.repositorios;
 import co.edu.uniquindio.uniMarket.entidades.Favorito;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface FavoritoRepo extends JpaRepository<Favorito, Long> {
+public interface FavoritoRepo extends JpaRepository<Favorito, Integer> {
+
+
 }
