+++ persistencia/src/main/java/co/edu/uniquindio/uniMarket/repositorios/ImagenRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ImagenRepo extends JpaRepository<Imagen, Long> {
+public interface ImagenRepo extends JpaRepository<Imagen, Integer> {
 }
