+++ persistencia/src/main/java/co/edu/uniquindio/uniMarket/repositorios/Producto_ModeradorRepo.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Producto_ModeradorRepo extends JpaRepository<Producto_Moderador, Long> {
+public interface Producto_ModeradorRepo extends JpaRepository<Producto_Moderador, Integer> {
 }
